# Indexes and caches in usvm-ts-dataflow