--- a/usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/annotation/TypeScheme.kt
+++ b/usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/annotation/TypeScheme.kt
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2022 UnitTestBot contributors (utbot.org)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.usvm.dataflow.ts.infer.annotation
+
+import org.jacodb.ets.base.EtsType
+import org.jacodb.ets.model.EtsMethod
+import org.usvm.dataflow.ts.infer.AccessPathBase
+
+interface MethodTypeScheme {
+    fun typeOf(base: AccessPathBase): EtsType?
+}
+
+interface TypeScheme {
+    fun methodScheme(method: EtsMethod): MethodTypeScheme
+    fun methodSchemes(): List<MethodTypeScheme>
+    fun thisType(method: EtsMethod): EtsType?
+}