… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/TypeInferenceManager.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/EntityId.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/VerificationResult.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/collectors/MethodSummaryCollector.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/collectors/StmtSummaryCollector.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/collectors/SummaryCollector.kt
… usvm-ts-dataflow/src/main/kotlin/org/usvm/dataflow/ts/infer/verify/collectors/Utils.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/EtsTypeAnnotationTest.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/EtsTypeInferenceTest.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/EtsTypeResolverAbcTest.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/EtsTypeResolverPerformanceTest.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/utils/AbcProjects.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/utils/PerformanceReport.kt
… usvm-ts-dataflow/src/test/kotlin/org/usvm/dataflow/ts/test/utils/TypeInferenceStatistics.kt