This PR improves TS dataflow engine.

# Performance report

## Baseline

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|6.5s|10.2s|14.2|
|project2|10.0s|10.6s|4.4|
|project3|3.3s|3.6s|7.3|
|project4|2.2s|2.5s|15.8|
|project5|2m 40s|3m 13s|11.4|
|project6|14.8s|21.0s|15.0|
|project7|2.7s|3.2s|10.8|

## With optimizations

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|4.0s|5.2s|13.9|
|project2|4.2s|5.4s|4.5|
|project3|1.5s|2.1s|7.3|
|project4|0.8s|1.1s|15.6|
|project5|1m 9s|1m 15s|11.3|
|project6|7.3s|7.9s|15.1|
|project7|0.8ms|1.1s|11.2|


## Add indexes for EtsApplicationGraph

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|4.2s|7.4s|14.0|
|project2|6.1s|6.2s|4.5|
|project3|2.1s|2.4s|7.3|
|project4|1.0s|1.3s|15.9|
|project5|1m 56s|2m 6s|11.4|
|project6|9.1s|10.2s|15.0|
|project7|1.1s|1.4s|10.8|

## Add cache for TypeGuesser

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|5.8s|7.6s|13.9|
|project2|9.6s|10.5s|4.5|
|project3|3.1s|3.3s|7.3|
|project4|2.1s|2.4s|16.2|
|project5|2m 3s|2m 34s|11.3|
|project6|18.2s|25.1s|15.0|
|project7|3.2s|4.2s|10.9|

## Add indexes for refinement

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|6.4s|9.1s|14.2|
|project2|9.3s|9.9s|4.5|
|project3|3.2s|3.5s|7.3|
|project4|2.1s|2.4s|16.3|
|project5|2m 22s|2m 45s|11.4|
|project6|14.5s|17.3s|15.4|
|project7|2.6s|3.0s|10.9|

## Optimized alias analysis

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|6.6s|9.4s|13.5|
|project2|9.6s|10.9s|4.6|
|project3|3.2s|4.2s|7.3|
|project4|2.4s|3.1s|16.2|
|project5|2m 20s|3m 2s|11.4|
|project6|14.8s|16.2s|15.0|
|project7|2.7s|3.1s|10.9|

## Add indexes to TypeFactProcessor

|project|min time|max time|%|
|:--|:--|:--|:--|
|project1|6.2s|8.4s|14.2|
|project2|9.6s|10.1s|4.5|
|project3|3.3s|3.8s|7.3|
|project4|2.2s|2.7s|16.3|
|project5|2m 29s|2m 57s|11.5|
|project6|15.4s|18.5s|15.2|
|project7|2.9s|4.0s|11.1|

# Other changes
* Removed unused `EntityId`
* tests: moved project processing into a separate method
* tests: added performance tests
* cleanup `TypeInferenceManager`